EA-9267 - Less copy and paste on things, make a base*.tsx file
