+++ pages/submit-a-ticket.tsx
@@ -3,7 +3,7 @@ import { getSupportNavigation } from 'lib/api';
 import { Box, Text, Button, Stack } from '@sparkpost/matchbox';
 import SEO from 'components/site/seo';
 import DocsLayout from 'components/site/docsLayout';
-import type { NavigationItemProps } from 'components/site/navigation';
+import type { NavigationItemProps } from 'components/site/baseNavigation';
 
 type SubmitATicketPageProps = {
   navigationData?: NavigationItemProps[];
